requirements.txt
playwright
tqdm
pillow
